% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightedLowess-package.R
\docType{package}
\name{WeightedLowess.tests-package}
\alias{WeightedLowess.tests}
\alias{WeightedLowess.tests-package}
\title{WeightedLowess.tests: R-based Tests for the WeightedLowess C++ code}
\description{
Some extra tests for the WeightedLowess code.
}
\author{
\strong{Maintainer}: Aaron Lun \email{infinite.monkeys.with.keyboards@gmail.com}

}
